Hello son

+++ CPP00/ex00/megaphone.cpp
@@ -11,7 +11,7 @@ int main(int argc, char **argv)
     for (int i = 1; i < argc; i++)
     {
         std::string str = argv[i];
-        for (int j = 0; j < str.length(); j++)
+        for (size_t j = 0; j < str.length(); j++)
             std::cout << (char) (str[j] ^ ((str[j] >= 'a' && str[j] <= 'z') << 5));
     }
     std::cout << '\n';
